JS3D: Return double instead of Integer for Queue Output getQueueLength
This change allows the queue length to be graphed

+++ src/main/java/com/sandwell/JavaSimulation3D/Queue.java
@@ -327,7 +327,7 @@ public class Queue extends DisplayEntity {
 	@Output(name = "QueueLength",
 	 description = "The present number of entities in the queue.",
 	    unitType = DimensionlessUnit.class)
-	public Integer getQueueLength(double simTime) {
+	public double getQueueLength(double simTime) {
 		return itemList.size();
 	}
